Merge remote-tracking branch 'origin/feat-mavlink-v2' into connection-mavlink-tests
The work on this branch is pushed before merging all the changes into master because it is crucial for the daedalus2-project to have a working connection in a few days which means this linking of all files must be pushed before the official "release" of the features.

This however also means that this branch will be shut down as soon as the tests are completed and will probably never see the main/master-branch. Use for at own risk!